+++ packages/framework/src/Console/Commands/PublishConfigsCommand.php
@@ -23,13 +23,29 @@ class PublishConfigsCommand extends Command
 
     public function handle(): int
     {
+        $options = [
+            'All configs',
+            '<comment>hyde-configs</comment>: Main configuration files',
+            '<comment>support-configs</comment>: Laravel and package configuration files',
+        ];
+        $selection = $this->choice('Which configuration files do you want to publish?', $options, 'All configs');
+
+        $tag = $this->parseTagFromSelection($selection, $options);
+
         Artisan::call('vendor:publish', [
-            '--tag' => 'configs',
+            '--tag' => $tag,
             '--force' => true,
         ], $this->output);
 
         $this->infoComment(sprintf('Published config files to [%s]', Hyde::path('config')));
 
         return Command::SUCCESS;
+    }
+
+    protected function parseTagFromSelection(string $selection, array $options): string
+    {
+        $tags = ['configs', 'hyde-configs', 'support-configs'];
+
+        return $tags[array_search($selection, $options)];
     }
 }