+++ packages/framework/src/Console/ConsoleServiceProvider.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Console;
 
+use Illuminate\Console\Application as Artisan;
 use Illuminate\Support\ServiceProvider;
 
 /**
@@ -36,6 +37,24 @@ class ConsoleServiceProvider extends ServiceProvider
 
             Commands\ChangeSourceDirectoryCommand::class,
         ]);
+
+        Artisan::starting(function (Artisan $artisan): void {
+            $artisan->setName(self::logo());
+        });
+    }
+
+    protected static function logo(): string
+    {
+        return <<<ASCII
+        
+        \033[34m     __ __        __   \033[33m ___  __ _____
+        \033[34m    / // /_ _____/ /__ \033[33m/ _ \/ // / _ \
+        \033[34m   / _  / // / _  / -_)\033[33m ___/ _  / ___/
+        \033[34m  /_//_/\_, /\_,_/\__/\033[33m_/  /_//_/_/
+        \033[34m       /___/
+            
+        \033[0m
+        ASCII;
     }
 
     public function boot(): void