+++ packages/framework/tests/Feature/Commands/PublishConfigsCommandTest.php
@@ -33,6 +33,7 @@ class PublishConfigsCommandTest extends TestCase
     public function test_command_has_expected_output()
     {
         $this->artisan('publish:configs')
+            ->expectsChoice('Which configuration files do you want to publish?', 'All configs', $this->expectedOptions())
             ->expectsOutput(sprintf('Published config files to [%s]', Hyde::path('config')))
             ->assertExitCode(0);
     }
@@ -41,7 +42,9 @@ class PublishConfigsCommandTest extends TestCase
     {
         $this->assertDirectoryDoesNotExist(Hyde::path('config'));
 
-        $this->artisan('publish:configs')->assertExitCode(0);
+        $this->artisan('publish:configs')
+            ->expectsChoice('Which configuration files do you want to publish?', 'All configs', $this->expectedOptions())
+            ->assertExitCode(0);
 
         $this->assertFileEquals(Hyde::vendorPath('config/hyde.php'), Hyde::path('config/hyde.php'));
 
@@ -53,8 +56,19 @@ class PublishConfigsCommandTest extends TestCase
         File::makeDirectory(Hyde::path('config'));
         File::put(Hyde::path('config/hyde.php'), 'foo');
 
-        $this->artisan('publish:configs')->assertExitCode(0);
+        $this->artisan('publish:configs')
+            ->expectsChoice('Which configuration files do you want to publish?', 'All configs', $this->expectedOptions())
+            ->assertExitCode(0);
 
         $this->assertNotEquals('foo', File::get(Hyde::path('config/hyde.php')));
+    }
+
+    protected function expectedOptions(): array
+    {
+        return [
+            'All configs',
+            '<comment>hyde-configs</comment>: Main configuration files',
+            '<comment>support-configs</comment>: Laravel and package configuration files',
+        ];
     }
 }