+++ RELEASE_NOTES.md
@@ -43,8 +43,11 @@ These are changes that break backwards compatibility and that are likely to conc
 
 - HydePHP now requires PHP 8.1 or higher.
 - Almost all namespaces in the framework have been changed and restructured.
+- Many of the internal underling Laravel application files have been moved, this means your `app/` directory will look a bit different. See [`#873`](https://github.com/hydephp/develop/pull/873)
+- The `config/site.php` file has been merged into `config/hyde.php`. See [`#964`](https://github.com/hydephp/develop/pull/964) for the upgrade guide.
 
 #### Breaking internal changes
+
 These are changes that break backwards compatibility but are unlikely to concern users using HydePHP to create sites.
 Instead, these changes will likely only concern those who write custom code and integrations using the HydePHP framework.